+++ src/coreclr/minipal/Windows/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(SOURCES
     doublemapping.cpp
     dn-u16.cpp
-    ${CLR_SRC_NATIVE_DIR}/minipal/is_native_debugger_present.c
+    ${CLR_SRC_NATIVE_DIR}/minipal/debugger.c
     ${CLR_SRC_NATIVE_DIR}/minipal/utf8.c
     ${CLR_SRC_NATIVE_DIR}/minipal/time.c
 )
