+++ src/coreclr/nativeaot/Runtime/CMakeLists.txt
@@ -51,7 +51,7 @@ set(COMMON_RUNTIME_SOURCES
     ${GC_DIR}/softwarewritewatch.cpp
 
     ${CLR_SRC_NATIVE_DIR}/minipal/cpufeatures.c
-    ${CLR_SRC_NATIVE_DIR}/minipal/is_native_debugger_present.c
+    ${CLR_SRC_NATIVE_DIR}/minipal/debugger.c
     ${CLR_SRC_NATIVE_DIR}/minipal/time.c
 )
