+++ src/coreclr/pal/src/CMakeLists.txt
@@ -252,8 +252,8 @@ if(CLR_CMAKE_TARGET_LINUX)
   add_library(tracepointprovider OBJECT misc/tracepointprovider.cpp)
 endif(CLR_CMAKE_TARGET_LINUX)
 
-add_library(is_native_debugger_present STATIC ${CLR_SRC_NATIVE_DIR}/minipal/is_native_debugger_present.c)
-target_link_libraries(coreclrpal PRIVATE is_native_debugger_present)
+add_library(minipal_debugger STATIC ${CLR_SRC_NATIVE_DIR}/minipal/debugger.c)
+target_link_libraries(coreclrpal PRIVATE minipal_debugger)
 
 if(CLR_CMAKE_TARGET_OSX)
   find_library(COREFOUNDATION CoreFoundation)
