+++ src/coreclr/utilcode/debug.cpp
@@ -12,7 +12,7 @@
 #include "ex.h"
 #include "corexcep.h"
 
-#include <minipal/is_native_debugger_present.h>
+#include <minipal/debugger.h>
 
 #ifdef _DEBUG
 #define LOGGING
