+++ src/mono/mono/utils/CMakeLists.txt
@@ -177,7 +177,7 @@ set(utils_common_sources
     wasm-module-reader.c
     mono-bitutils.h
     mono-bitutils.c
-    ${CLR_SRC_NATIVE_DIR}/minipal/is_native_debugger_present.c
+    ${CLR_SRC_NATIVE_DIR}/minipal/debugger.c
   )
 
 if(MONO_CROSS_COMPILE)
